2.2.0.0
Changed the main task of the plugin, now it is Hiding links from guests.
Ability to disable referer hiding for links without disabling other plugin functions.
The fix from this post has been moved to the main update branch.
Extension of plugin support to xenforo 2.2.x version
Minor fixes in master language phrases.
Services for hiding referer on links that do not work and do not support https have been removed from the plugin.

Upgrading is optional if you are satisfied with the features of previous versions.